requirements.txt
diffusers 
transformers 
torch
accelerate
customtkinter
python-dotenv